make single gear function more flexible
