+++ wheels/wheels.go
@@ -2,7 +2,6 @@ package wheels
 
 import (
 	"fmt"
-	"test_code/motor"
 )
 
 type Wheels struct {
@@ -17,7 +16,6 @@ type WheelsManipulator interface {
 	GetRPM() int
 	GetSize() int
 	GetCount() int
-	SetHalfMotorRPM(motor.MotorManipulator) error
 	Print()
 }
 
@@ -26,16 +24,6 @@ func (w *Wheels) SetRPM(rpm int) error {
 		return fmt.Errorf("Negative RPM")
 	}
 	w.Rpm = rpm
-	return nil
-}
-
-func (w *Wheels) SetHalfMotorRPM(m motor.MotorManipulator) error {
-	motorRpm := m.GetRPM()
-	rpm := int(motorRpm / 2)
-	err := w.SetRPM(rpm)
-	if err != nil {
-		return err
-	}
 	return nil
 }
